+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 - Fixed WebInbox Runtime Issues
 - Skipped WebInbox Retries when config not present
 - Native Display KV_Pair Custom Event Paylaod Backward Compatibility
+- Enhanced session storage check for WZRK_D with fallback and error handling.
+
 
 ## [1.13.4] 7th Mar, 2025
 - Added support to preview html for custom native html